benchmarking/pyproject.toml
[project]
name = "benchmarking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groundlight>=0.26.0",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
]
